requirements.txt
openai==1.12.0
pandas==3.0.0
numpy==1.26.3
scikit-learn==1.4.0
chromadb==0.4.22
python-dotenv==1.0.0
jupyter==1.0.0
fastapi==0.109.0
uvicorn==0.27.0